Renamed the NodeStartupSpec to NodeJoinAndUpSpec and added tests for both JOINING and UP.
